refactor: article 테이블 에 맞게 Article 클래스 수정
- id 를 Integer 타입에서 Long 타입으로 변경하였습니다.

+++ src/main/java/com/kakao/cafe/article/controller/ArticleController.java
@@ -55,7 +55,7 @@ public class ArticleController {
     }
 
     @GetMapping("/articles/{id}")
-    public ModelAndView viewArticle(@PathVariable("id") int id, ModelAndView modelAndView) {
+    public ModelAndView viewArticle(@PathVariable("id") Long id, ModelAndView modelAndView) {
         Article article = articleService.viewArticle(id);
 
         modelAndView.setViewName("qna/show");
